Create healf barz.png
concept art for "wave hp", basically as a timer for how long you have left. it would go down depending on how many aliens you've killed (featured are 100% of aliens are still alive, 50%, and 25%. they're supposed to look like green goo in a jar that turns more red as the wave goes on.